Fixes request for ExistKeyValueAsync
